feat: 用户体系与双令牌登录、标签拖拽排序、编排页选择器增强、图标与关于页
用户与权限（auth）
- users 增 is_admin / last_login_at，新增 user_sessions 表（access 2h + refresh 30d 滑动续期，
  只存 sha256，每次刷新轮换 refresh）
- 首个注册用户即管理员（判据为「不存在可登录的管理员」，兼容库中无密码占位行）
- 错误码分层：61=未登录/无权限，4401=登录态过期，前端据此自动续期并重试原请求
- 中间件只识别身份不拦截，权限点在 logic（MustLogin / MustAdmin）
- 前端 /login 页 + useAuth；未登录一律只读，写入口（新建/编辑/收藏/编排/保存）全部隐藏

标签
- 新增 PUT /tags/sort（tag_reorder），设置页拖拽 ⠿ 调整顺序并落库
- manifest/init.sql 移除预置标签 INSERT，避免重跑脚本覆盖现有数据

前端
- 编排页选择器：同一餐次/一周内菜谱去重，「顿顿不重样」开关持久化到本地，收藏夹可作为来源
- 文案里的文字「+」统一换成 PlusIcon 组件；设置页新增「关于」页签
- README / AGENTS.md 同步以上变更（含 auth 权限矩阵与 operationId reorder 说明）

+++ internal/model/entity/attachments.go
@@ -12,7 +12,7 @@ type Attachments struct {
 	FileName    string `json:"fileName"    orm:"file_name"     description:""` //
 	StoragePath string `json:"storagePath" orm:"storage_path"  description:""` //
 	MimeType    string `json:"mimeType"    orm:"mime_type"     description:""` //
-	SizeBytes   int    `json:"sizeBytes"   orm:"size_bytes"    description:""` //
+	SizeBytes   int64  `json:"sizeBytes"   orm:"size_bytes"    description:""` //
 	Width       int    `json:"width"       orm:"width"         description:""` //
 	Height      int    `json:"height"      orm:"height"        description:""` //
 	Duration    int    `json:"duration"    orm:"duration"      description:""` //
